Model is very close to working.
I found one major error in the definition of the rear wheel center velocity. I
was using the same definition from my dissertation which assumed that no is the
origin and dn is moving wrt to no through q1 and q2. But I think, not quite
sure, that when you set the rear wheel contact point as the origin you have to
account for the velocity components due to rear wheel rate and pitch rate.
Making that change, which matches Gilbert's version, gets the ode eval very
close to Basu-Mandal numbers. I just have a small error at this point. I'm also
still getting a singular matrix on the ode function generator (this may be a
larger problem or misunderstanding as I see it too often in other problems).